Update to django 1.6
