+++ =30
@@ -1,2 +1,2 @@
-Requirement already up-to-date: pip in ./.env/lib/python2.7/site-packages (20.0.2)
-Requirement already up-to-date: setuptools in ./.env/lib/python2.7/site-packages (44.0.0)
+Requirement already satisfied: pip in ./.env/lib/python2.7/site-packages (20.0.2)
+Requirement already satisfied: setuptools in ./.env/lib/python2.7/site-packages (44.1.0)
